Complete Refactor of the S3 Importers
