--- spec/fixtures/files/wonderschool_necc_dashboard_data_invalid_format.csv
@@ -1,2 +1,0 @@
-wrong_headers,icon,face
-nope,yep,maybe
